Rename package to org.jgroups.protocols.azure for future compatibility reasons
