+++ azure/src/main/java/org/jgroups/protocols/azure/AZURE_PING.java
@@ -1,4 +1,4 @@
-package org.jgroups.azure.protocols;
+package org.jgroups.protocols.azure;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
